// src/app/api/model/tvEpisode.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GuestStar } from './guestStar';
import { CrewMember } from './crewMember';


export interface TvEpisode { 
    air_date?: string;
    crew?: Array<CrewMember>;
    episode_number?: number;
    guest_stars?: Array<GuestStar>;
    name?: string;
    overview?: string;
    id?: number;
    production_code?: string;
    runtime?: number | null;
    season_number?: number;
    still_path?: string | null;
    vote_average?: number;
    vote_count?: number;
}